requirements.txt
streamlit
opencv-python-headless
mediapipe-silicon
numpy
streamlit-webrtc